feat: print value name and its value
