+++ tools/sanity-check.py
@@ -138,25 +138,6 @@ def sanity_check(directories):
     return errors
 
 
-def check_commit_message():
-    print(os.environ['CHANGE_TITLE'])
-    print(os.environ['CHANGE_URL'])
-    print('CHANGE_AUTHOR=%s' % os.environ['CHANGE_AUTHOR'])
-    print('CHANGE_AUTHOR_DISPLAY_NAME=%s' % os.environ['CHANGE_AUTHOR_DISPLAY_NAME'])
-    print('CHANGE_AUTHOR_EMAIL=%s' % os.environ['CHANGE_AUTHOR_EMAIL'])
-    print('CHANGE_BRANCH=%s' % os.environ['CHANGE_BRANCH'])
-    print('CHANGE_TARGET=%s' % os.environ['CHANGE_TARGET'])
-    print('CHANGE_ID=%s' % os.environ['CHANGE_ID'])
-    print('BRANCH_NAME=%s' % os.environ['BRANCH_NAME'])
-    if 'CHANGE_FORK' in os.environ:
-        print("CHANGE_FORK=%s" % os.environ['CHANGE_FORK'])
-    if 'TAG_NAME' in os.environ:
-        print(os.environ['TAG_NAME'])
-        print(os.environ['TAG_TIMESTAMP'])
-        print(os.environ['TAG_UNIXTIME'])
-        print(os.environ['TAG_UNIXTIME'])
-
-
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
         description='check consistency between directory structure, pom.xml and build.properties')
@@ -165,7 +146,6 @@ if __name__ == '__main__':
                         default='main/plugins:test/plugins')
     args = parser.parse_args()
     plugin_directories = args.directories.split(':')
-    check_commit_message()
     nb_errors = sanity_check(plugin_directories)
     if nb_errors != 0:
         print("%sProject is not correctly configured (%s %s found).%s" % (
